viral/castor_krfe/requirements.txt
numpy
scipy
scikit-learn
biopython
